+++ client/src/environments/environment.production.ts
@@ -1,6 +1,6 @@
 export const environment = {
   production: true,
-  version: 'v0.7.0-beta',
+  version: 'v0.8.0-beta',
   assetUrl: '/static',
   apiUrl: '/api',
   wsUrl: '/ws',